Use Docker Compose to provide credential secret
The actual secret is in a file in secrets/ which is added to .gitignore
to keep it out of version control. The secret should never be checked
in.

DC provides the secret to the container at runtime as a file, and
settings.js now has provisions to read that file (or report an error).